edit: move mysql Error enum to mysql::error.rs
