+++ server/src/mysql/mod.rs
@@ -1,9 +1,10 @@
 // Copyright 2022 CeresDB Project Authors. Licensed under Apache-2.0.
 
 mod builder;
+pub mod error;
 mod handler;
 mod worker;
 mod writer;
 
-pub use builder::{Builder, Config as MysqlConfig, Error as MysqlBuilderError};
-pub use handler::{Error as MysqlHandlerError, MysqlHandler};
+pub use builder::{Builder, Config as MysqlConfig};
+pub use handler::MysqlHandler;
